Inline tiny model/sticker/ modules
All other top-level model modules don't have such teeny tiny sub modules
